feat(anim): soportar 60 fotogramas a 15fps
- MAX_ANIM_FRAMES de 20 a 60
- animFramesBitmap de uint32_t a uint64_t (60 slots no caben en 32 bits),
  shifts con 1ULL
- Timeout de descarga de 5s a 15s (60 frames ~490KB tardan ~12s por MQTT)

Validado en hardware v2 (ESP32-S3): descarga 59/59 frames en ~12s y
reproduce a 15fps. Los v1 sin PSRAM siguen recortando y submuestreando
en runtime como hasta ahora.

+++ src/globals.cpp
@@ -114,7 +114,7 @@ unsigned long animLastFrameTime = 0;
 unsigned long animLoopCount = 0;
 uint8_t animFrameStep = 1;
 unsigned long animFrameInterval = 200;
-uint32_t animFramesBitmap = 0;
+uint64_t animFramesBitmap = 0;
 unsigned long animDownloadStartTime = 0;
 uint8_t animRetryCount = 0;
